route getbypricerange

+++ src/api/clothes/clothes.repository.js
@@ -3,8 +3,8 @@ import clothesModel from './clothes.model.js';
 function getByPriceRange({ query, }) {
   const price = { $gte: 0, };
 
-  query.min && (price.$gte = parseInt(query.min));
-  query.max && (price.$lte = parseInt(query.max));
+  query.min && (price.$gte = parseFloat(query.min));
+  query.max && (price.$lte = parseFloat(query.max));
 
   const filteredByPriceRangeClothes = clothesModel.find({ price, }).lean();
   return filteredByPriceRangeClothes;
